Draggable

+++ BroadsideRepo/Broadside/src/com/starboardstudios/broadside/interfaces/Draggable.java
@@ -9,10 +9,10 @@ public interface Draggable
     public abstract void dragStarted();
 
 
-    public abstract void midDrag(int x,int y);
+    public abstract void midDrag(float x,float y);
 
 
-    public abstract void endDrag(int x, int y);
+    public abstract void endDrag(float x, float y);
